# openapi.yaml
openapi: 3.0.3
info:
  title: LogLynx API
  description: |
    LogLynx is an advanced log analytics platform for analyzing Traefik access logs and other log sources.
    This API provides comprehensive statistics, real-time metrics, and detailed insights into your web traffic.

    ## Features
    - **Summary Statistics**: Get overall metrics including requests, visitors, bandwidth, and response times
    - **Timeline Data**: Analyze traffic patterns over time with customizable time ranges
    - **Top Statistics**: Identify top paths, countries, IPs, user agents, and more
    - **Distribution Analytics**: View distributions of status codes, HTTP methods, protocols, and device types
    - **Real-time Monitoring**: Stream live metrics via Server-Sent Events (SSE)
    - **Performance Metrics**: Track response times and backend health
    - **Geographic Analytics**: Analyze traffic by country and region
    - **IP Analytics**: Deep-dive analytics for specific IP addresses
    - **IP Tagging**: Assign friendly names and custom tags to IP addresses for better organization (Available from v1.1.1)
    - **Service Filtering**: Filter by backend name, URL, or host with intelligent auto-detection

    ## Service Filtering
    Most endpoints support service-based filtering with the following parameters:

    ### Single Service Filtering
    - `service`: Service name to filter by (e.g., "api-service", "web-frontend")
...
      - `backend_url`: Filter by backend URL
...

...
servers:
...
paths:
  /stats/summary:
    get:
      tags:
        - Summary
      summary: Get summary statistics
      description: |
        Returns overall statistics including total requests, unique visitors, bandwidth, success rates, and top items.

        **Service Filtering**: Use `service` and `service_type` for single service filtering, or `services[]` and `service_types[]` for multi-service filtering.
        Legacy `host` parameter is supported for backward compatibility.

        **Hide My Traffic**: Use `exclude_own_ip=true` to exclude your IP from results.
      operationId: getSummary
      parameters:
        - $ref: '#/components/parameters/ServiceFilter'
        - $ref: '#/components/parameters/ServiceTypeFilter'
        - $ref: '#/components/parameters/ServicesArray'
        - $ref: '#/components/parameters/ServiceTypesArray'
        - $ref: '#/components/parameters/HostFilter'
        - $ref: '#/components/parameters/HoursParam'
        - $ref: '#/components/parameters/ExcludeOwnIP'
        - $ref: '#/components/parameters/ExcludedIPs'
        - $ref: '#/components/parameters/ExcludeServices'
        - $ref: '#/components/parameters/ExcludeServiceTypes'
      responses:
        '200':
          description: Summary statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsSummary'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /widget/data:
    get:
      tags:
        - Widget
      summary: Get real-time widget data
      description: |
        Returns compact real-time metrics optimized for widget embedding.
        Shows data from the last hour with requests per minute calculation.

        **Note**: This endpoint is only available when `WIDGET_ENABLED=true` (default).
      operationId: getWidgetData
      responses:
        '200':
          description: Real-time widget data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WidgetRealtimeData'
        '404':
          description: Widget endpoints disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /widget/summary:
    get:
      tags:
        - Widget
      summary: Get widget summary data for a time period
      description: |
        Returns compact summary statistics for embedding in widgets.
        Optimized for displaying aggregated data over a time period.

        **Note**: This endpoint is only available when `WIDGET_ENABLED=true` (default).
      operationId: getWidgetSummary
      parameters:
        - name: hours
          in: query
          description: Number of hours to analyze (1-720 for up to 30 days, default 24)
          schema:
            type: integer
            minimum: 1
            maximum: 720
            default: 24
            example: 24
      responses:
        '200':
          description: Widget summary data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WidgetSummaryData'
        '404':
          description: Widget endpoints disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /widget/timeline:
    get:
      tags:
        - Widget
      summary: Get widget timeline data
      description: |
        Returns compact timeline data optimized for sparkline charts in widgets.
        Returns up to 30 data points for periods <= 24 hours, or 50 points for longer periods.

        **Note**: This endpoint is only available when `WIDGET_ENABLED=true` (default).
      operationId: getWidgetTimeline
      parameters:
        - name: hours
          in: query
          description: Number of hours to analyze (1-720 for up to 30 days, default 24)
          schema:
            type: integer
            minimum: 1
            maximum: 720
            default: 24
            example: 24
      responses:
        '200':
          description: Widget timeline data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WidgetTimelinePoint'
        '404':
          description: Widget endpoints disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /ip/tags:
    get:
      tags:
        - IP Tagging
      summary: Get all IP tags
      description: Returns a list of all assigned IP tags and friendly names (Available from v1.1.1)
      operationId: getAllTags
      responses:
        '200':
          description: List of all IP tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IPTag'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - IP Tagging
      summary: Create or update IP tag
      description: Assign a friendly name and tags to an IP address (Available from v1.1.1)
      operationId: createOrUpdateTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IPTagInput'
      responses:
        '200':
          description: Tag updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IPTag'
        '400':
          description: Invalid input
        '500':
          $ref: '#/components/responses/InternalServerError'
  /ip/tags/{ip}:
    get:
      tags:
        - IP Tagging
      summary: Get tags for a specific IP
      description: Returns friendly name and tags for the specified IP address (Available from v1.1.1)
      operationId: getTag
      parameters:
        - name: ip
          in: path
          description: IP address to retrieve tags for
          required: true
          schema:
            type: string
          example: 192.168.1.100
      responses:
        '200':
          description: IP tag details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IPTag'
        '404':
          description: IP tags not found
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - IP Tagging
      summary: Delete IP tags
      description: Removes all tags and friendly name assigned to the specified IP address (Available from v1.1.1)
      operationId: deleteTag
      parameters:
        - name: ip
          in: path
          description: IP address to remove tags from
          required: true
          schema:
            type: string
          example: 192.168.1.100
      responses:
        '200':
          description: Tags deleted successfully
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
...
    HostFilter:
      name: host
      in: query
      description: |
        **Legacy parameter** - Filter results by backend/service name.
        Equivalent to using `service` parameter with `service_type=auto`.
        Kept for backward compatibility with older clients.
      required: false
      schema:
        type: string
      example: my-service

    # New single service filtering
    ServiceFilter:
      name: service
      in: query
      description: Service name to filter by (backend name, URL, or host depending on service_type)
      required: false
      schema:
        type: string
      example: api-service
    ServiceTypeFilter:
      name: service_type
      in: query
      description: |
        Type of service identifier to use for filtering:
        - `auto` (default): Automatically detects and uses backend_name, backend_url, or host based on availability
        - `backend_name`: Filter specifically by Traefik backend name
        - `backend_url`: Filter by backend URL
        - `host`: Filter by HTTP Host header
      required: false
      schema:
        type: string
        enum: [auto, backend_name, backend_url, host]
        default: auto
      example: backend_name

    # Multi-service filtering (arrays)
    ServicesArray:
      name: services[]
      in: query
      description: |
        Array of service names for multi-service filtering.
        Must be used together with `service_types[]` parameter with matching array length.
        Can be repeated multiple times: `?services[]=api&services[]=web`
      required: false
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
      example: [api-service, web-frontend]
    ServiceTypesArray:
      name: service_types[]
      in: query
      description: |
        Array of service types corresponding to `services[]` parameter.
        Must have the same length as `services[]` array.
        Each element can be: auto, backend_name, backend_url, or host
      required: false
      schema:
        type: array
        items:
          type: string
          enum: [auto, backend_name, backend_url, host]
      style: form
      explode: true
      example: [backend_name, host]

    # Hide My Traffic parameters
    ExcludeOwnIP:
      name: exclude_own_ip
      in: query
      description: |
        Set to `true` to exclude requests from your own IP address.
        The server automatically detects your IP from the request.
        Use `excluded_ips[]` to exclude additional manually entered IP addresses.
        When `true` without `exclude_services[]`, excludes your IP from ALL services.
      required: false
      schema:
        type: boolean
        default: false
      example: true
    ExcludedIPs:
      name: excluded_ips[]
      in: query
      description: |
        Array of manually entered IP addresses to exclude from analytics.
        These IPs are combined with the detected client IP when `exclude_own_ip=true`.
        If `exclude_services[]` is omitted, all listed IPs are excluded from all services.
        If `exclude_services[]` is provided, all listed IPs are excluded only for those services.
      required: false
      schema:
        type: array
        items:
          type: string
          format: ip
      style: form
      explode: true
      example: [192.168.1.10, 2001:db8::1]
    ExcludeServices:
      name: exclude_services[]
      in: query
      description: |
        Array of service names to exclude your IP from (used with `exclude_own_ip=true`).
        If omitted, your IP is excluded from all services.
        If specified, your IP is only excluded from the listed services.
        Must be used together with `exclude_service_types[]`.
      required: false
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
      example: [admin-panel, internal-api]
    ExcludeServiceTypes:
      name: exclude_service_types[]
      in: query
      description: |
        Array of service types corresponding to `exclude_services[]`.
        Must have the same length as `exclude_services[]` array.
      required: false
      schema:
        type: array
        items:
          type: string
          enum: [auto, backend_name, backend_url, host]
      style: form
      explode: true
      example: [backend_name, host]

    # Common parameters
...
    HoursParam:
      name: hours
      in: query
      description: Number of hours to look back (1-8760 for up to 1 year, 0 for all time, default 168 for 7 days)
      schema:
        type: integer
        minimum: 0
        maximum: 8760
        default: 168
        example: 168
...
  schemas:
    StatsSummary:
      type: object
      properties:
        total_requests:
          type: integer
          format: int64
          description: Total number of requests
          example: 125643
        valid_requests:
          type: integer
          format: int64
          description: Number of successful requests (2xx, 3xx)
          example: 119234
        failed_requests:
          type: integer
          format: int64
          description: Number of failed requests (4xx, 5xx)
          example: 6409
        unique_visitors:
          type: integer
          format: int64
          description: Number of unique IP addresses
          example: 1523
        unique_files:
          type: integer
          format: int64
          description: Number of unique paths accessed
          example: 342
        unique_404:
          type: integer
          format: int64
          description: Number of unique 404 paths
          example: 23
        total_bandwidth:
          type: integer
          format: int64
          description: Total bandwidth in bytes
          example: 5368709120
        avg_response_time:
          type: number
          format: double
          description: Average response time in milliseconds
          example: 142.5
        success_rate:
          type: number
          format: double
          description: Success rate percentage (0-100)
          example: 94.9
        not_found_rate:
          type: number
          format: double
          description: 404 error rate percentage
          example: 3.2
        server_error_rate:
          type: number
          format: double
          description: 5xx error rate percentage
          example: 1.9
        requests_per_hour:
          type: number
          format: double
          description: Average requests per hour
          example: 745.3
        top_country:
          type: string
          description: Country code with most traffic
          example: "US"
        top_path:
          type: string
          description: Most accessed path
          example: "/api/v1/users"
...
    IPTag:
      type: object
      description: Tag information for an IP address
      properties:
        id:
          type: integer
          example: 1
        ip_address:
          type: string
          example: "192.168.1.100"
        friendly_name:
          type: string
          example: "Office Server"
        tags:
          type: string
          example: "production,internal"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    IPTagInput:
      type: object
      required:
        - ip_address
      properties:
        ip_address:
          type: string
          example: "192.168.1.100"
        friendly_name:
          type: string
          example: "Office Server"
        tags:
          type: string
          example: "production,internal"
    WidgetRealtimeData:
      type: object
      description: Compact real-time metrics for widget embedding
      properties:
        status:
          type: string
          description: System health status
          enum: [healthy, warning, danger, error]
          example: "healthy"
        requests_per_minute:
          type: number
          format: double
          description: Requests per minute (calculated from last hour)
          example: 125.5
        error_rate:
          type: number
          format: double
          description: Combined error rate percentage (4xx + 5xx)
          example: 2.5
        avg_response_time:
          type: number
          format: double
          description: Average response time in milliseconds
          example: 145.3
        unique_ips:
          type: integer
          format: int64
          description: Number of unique IP addresses in the last hour
          example: 1523
    WidgetSummaryData:
      type: object
      description: Compact summary statistics for widget embedding
      properties:
        status:
          type: string
          description: System health status
          enum: [healthy, warning, danger, error]
          example: "healthy"
        total_requests:
          type: integer
          format: int64
          description: Total number of requests in the time period
          example: 125643
        requests_per_hr:
          type: number
          format: double
          description: Average requests per hour
          example: 5235.1
        error_rate:
          type: number
          format: double
          description: Combined error rate percentage (4xx + 5xx)
          example: 2.5
        avg_response_ms:
          type: number
          format: double
          description: Average response time in milliseconds
          example: 145.3
        unique_ips:
          type: integer
          format: int64
          description: Number of unique IP addresses
          example: 15234
        bandwidth_mb:
          type: number
          format: double
          description: Total bandwidth in megabytes
          example: 5123.5
    WidgetTimelinePoint:
      type: object
      description: Single point in widget timeline data
      properties:
        hour:
          type: string
          format: date-time
          description: Timestamp of this data point
          example: "2025-11-03T14:00:00Z"
        requests:
          type: integer
          format: int64
          description: Number of requests in this time bucket
          example: 1523
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Failed to retrieve data"
...
  responses:
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'